+++ packages/flame/CHANGELOG.md
@@ -41,6 +41,7 @@
  - Update `Camera` docs to showcase usage with `Game` class
  - Fixed a bug with `worldBounds` being set to `null` in `Camera`
  - `MockCanvas` is now strongly typed and matches numeric coordinates up to a tolerance
+ - Add `loadAllImages` to `Images`, which loads all images from the prefixed path
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps